pyproject.toml
[tool.poetry]
name = "rtdjango"
version = "0.1.0"
description = ""
authors = ["Gabriel Flores <gabriel.flores0503@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.3"
djangorestframework = "^3.11.0"
channels = "^2.4.0"
channels-redis = "^2.4.1"
psycopg2 = "^2.8.4"

[tool.poetry.dev-dependencies]
pytest = "^3.4"